chore(gql types): renames the source file PE-1337
